This pr is not ready to merge. I just want to show something.
A DNF is defined as:
```
(x_1 & x_2 & x_3) | (!x_1 & x_2 & x_3) | (x_1 & !x_2 & !x_3)
```
I found that build a BDD of DNF is quite time-wasting(maybe I do something wrong).
So I write a recursive way to deal with it. It visit every atom only once.

```
new                     time:   [31.313 ms 31.428 ms 31.562 ms]                
Found 13 outliers among 100 measurements (13.00%)
  8 (8.00%) high mild
  5 (5.00%) high severe

Benchmarking old: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 145.8s, or reduce sample count to 10.
old                     time:   [1.4519 s 1.4528 s 1.4538 s]                   
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```
## Some questions
1. If there is a better way to DNF already in biodivine-lib-bdd or in other implementation?
2. If DNF is a universal circumstance? At least for me, it's the only input format.
3. `Bitset` instead of `Vec<Bool>`?
4. CNF

Reference:
 https://www.cmi.ac.in/~madhavan/courses/verification-2011/andersen-bdd.pdf  exercise 4.3