@daemontus 2021-08-10
This looks like it could be a nice addition at some point!

At the moment, what you implemented is probably the best we can do in `lib-bdd`. I also don't see any DNF or clause-specific API in other libraries (e.g. [CUDD](https://add-lib.scce.info/assets/doxygen-cudd-documentation/cudd_8h.html)), but I may be missing something. So as far as I can tell, people generally don't target this specifically in implementations. But that doesn't mean we can't have a special method for it.

However, in this particular case, the design of `lib-bdd` is indeed quite inefficient, because there is a lot of rather trivial operations (especially the inner for loop). The BDDs still need to be allocated/deallocated for each operation, hence it takes so long. We have an experimental API that adds native operations for "cubes" (essentially clauses in DNF) - see branch `cubes`. This should speed things up quite a bit once released.

I have merged everything into a single branch [here](https://github.com/sybila/biodivine-lib-bdd/blob/e5411f50a25c2bf71a1bf1028311c2d9f4802690/benches/dnf.rs) and added a comparison with CUDD as a "baseline". With `cubes`, the operation is still slower than your specialised implementation, but at least it is not as "tragic" :D

```
new                     time:   [70.356 ms 71.310 ms 72.283 ms]     
cudd                    time:   [227.98 ms 235.88 ms 244.69 ms]    
old-cubes               time:   [202.68 ms 211.57 ms 223.36 ms]  
old                     time:   [2.3704 s 2.4167 s 2.4676 s] 
```

So as a short term solution, I can merge the cube API and publish it in a new version. If you still need better performance, we can build a production ready version of your DNF algorithm - I think it is a useful feature to have.

Misc:
 - Input format: We generally start with small, human-written formulas that are quick to import (the big ones are then built from that), so we generally don't care if the formula is CNF/DNF/etc. However, I wouldn't mind having a specialised methods for this.
 - `Vec<bool>` vs. `BitSet`: We can test that, but I don't expect it to make much of a difference. The advantage is mostly in smaller memory footprint, so if you can fit your DNF formula into memory with `Vec<bool>`, I would go with what's easier to use.