@daemontus 2021-08-10
I've created a pull request for the `cubes` API here: #20. I'll try to look into it soon :)