@daemontus 2021-09-01
In version [`0.3.0`](https://github.com/sybila/biodivine-lib-bdd/releases/tag/v0.3.0), there is now a new (more efficient) API for creating DNF/CNF formulas.

This is still not as efficient as this PR proposes (that would require lifting the `Bdd` node order requirement which has not happened yet). Hence I will not be closing this yet, however an issue (#24) has been created to track progress on this new implementation.