@daemontus 2023-05-03
It only took ... two years ... but the new CNF/DNF translation algorithms are now implemented in version `0.5.0`. 

If this is still (somewhat) relevant for you, please test if the new implementation performs as you would expect :)

Otherwise, I am closing this as no longer relevant.